added function to analysis.py
